+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 This is the log of notable changes to the Design System Nimbus that are developer-facing.
 Package-specific changes not released in any package will be added here just before the release. Until then, you can find them in changelogs of the individual packages (see [packages](./packages) directory).
 
-## 2025-08-05 `1.8.0`
+## 2025-08-13
+
+#### 🎉 New features
+
+- Added `@common` folder to `react` package to share common utilities between packages. ([#336](https://github.com/TiendaNube/nimbus-design-system/pull/336) by [@joacotornello](https://github.com/joacotornello))
+
+## 2025-08-05
 
 #### 🎉 New features